build: fix pkgconfig files (#1913)
### Description of changes: 

Currently the pkgconfig files are broken when aws-lc is built with Nix. 

See
https://github.com/jtojnar/cmake-snips?tab=readme-ov-file#concatenating-paths-when-building-pkg-config-files,
and the PR made to libfmt https://github.com/fmtlib/fmt/pull/1657

### Call-outs:

I'm not sure if bumping the minimum required CMake version is preferred.
The alternative would be to add
https://github.com/jtojnar/cmake-snips/blob/master/CMakeScripts/JoinPaths.cmake
to the `cmake/` directory in the repository. However, most linux
distributions have at least CMake 3.20, if not 3.25 so I don't see why
this would be a problem.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.