Support encode or decode ∞ like OpenSSL (#1930)
Ruby has tests asserting that OpenSSL encodes and decodes infinity
points to 0. This doesn't quite make sense, but our hand is sort of
forced here. It does show that there's dependence on this subtle
behavior and we might as well support it since we're trying to gain
better compatibility.

```
    assert_equal false, point.infinity?
    point.set_to_infinity!
    assert_equal true, point.infinity?
    assert_equal 0.to_bn, point.to_bn
    assert_equal B(%w{ 00 }), point.to_octet_string(:uncompressed)
    assert_equal true, point.on_curve?
```

### Call-outs:
Tweak against changes from these commits;
* d9e2702
* a7185da

### Testing:
Tweaked existing tests that checked we did not support this behavior. 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.