Add support for POINT_CONVERSION_HYBRID (#1936)
This is the first time it has come up, but Ruby depends on the
`POINT_CONVERSION_HYBRID` format in its tests and expects the underlying
libcrypto to support it. There's not too much a difference between
`POINT_CONVERSION_HYBRID` and
`POINT_CONVERSION_UNCOMPRESSED` except that the first bit contains
information about the quadratic equation y is at. We have the building
blocks to support and can get this working with some tweaking around.
Most of this change is working with the first byte
`POINT_CONVERSION_HYBRID` does differently and point the encoding to
the right places for parsing.

### Testing:
* Testing parsing from and back with `POINT_CONVERSION_HYBRID`. What
we're parsing to and from hybrid doesn't really matter here, they're
already covered by other code paths.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.