Add further SIMD load/store variants to the ARM decoder
This change enables the following in the ARM decoder

 * 128-bit LDR/STR with post-index

 * 128-bit LDP/STP (previously only 64-bit was allowed)

 * LDUR and STUR

All these map to existing instruction abstract syntax (LDUR/STUR is
treated the same as LDR/STR except for the offset encoding). Thus, the
main changes are in the decoder only, though some proof support is
added, duplicating for 128 bits the naive way in which the two memory
locations written by STP are proved orthogonal during simulation.