Add SIMD multiplication by a single element to the ARM model
This allows a variant of SIMD multiplication "by element", i.e.
multiplying each lane of the first source operand by the *same*
single lane of the second source operand.

To express this conveniently, we introduce "lane" components
LANE_B, LANE_H, LANE_S and LANE_D, so for example the ARM
assembler syntax "v2.h[1]" corresponds to "Q2 :> LANE_H 1"
in the abstract syntax. These lane components include appropriate
duplication to an equivalent SIMD argument, so that the core
operation is expressed uniformly for the "by element" and
"vector" variants.