Account for cipher auth with multiple cert slots (#1956)
Ruby has a dependency on the multiple certificate slot mechanisms that
OpenSSL allows for. We've already done the work to support this, but
another Ruby 3.1 test exposed a gap in our support. We were only
looking at the negotiated signature algorithms to retrieve the right
certificate for the server to send back, but the cipher authentication
scheme was also checked in OpenSSL as well. Ruby's tests happen to
only depend on configuring the authentication scheme which brought
this to light.
We already happen to do this when checking private keys for TLS 1.0/1.1,
this also fixes it to check the cipher against the initial public key.
This change also introduces all of the mentioned cipher authenticated
check behavior for TLS 1.2.

### Testing:
New unit test that allows all possible sigalgs, with the cipher
authentication suite being the only restriction.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.