Add Cyrus-SASL to our CI (#1988)
### Issues:
`CryptoAlg-2554`

### Description of changes: 
Add Cyrus-SASL to our CI

### Call-outs:
There is no general testing suite for Cyrus. To test the GSSAPI with
Kerberos, we can use provided sample_client and sample_server
applications. However, this requires a working Kerberos setup with a KDC
and another script to support communication between the sample_client
and server programs. This is not being added as of now.


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.