Revert "Replace CONF's internal representation with something more typesafe" (#1986)
This reverts commit cc932cf394d2af845fdc46eac14f21886d5c98c9.

OpenSSL's historically made the `CONF` struct non-opaque. This hasn't
changed in OpenSSL 3 either although there is documentation that
indicates they wish to do so in the long run. It's likely going to be
some time before this is actually done, but in the meantime Ruby has a
dependency on `CONF`'s internals.
This commit is reverting some of the original upstream work done that
changes the internal variables of `CONF`. There will be a follow up PR
to make the structure non-opaque.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.