Expose a bit of lhash/conf for Ruby (#1987)
This extends upon https://github.com/aws/aws-lc/pull/1986.

We do have all the necessary support for "LHASH", it's just been renamed
and made internal by upstream: b8441ac. Ruby's CONF module happens
to have a dependency on a couple of LHASH macros OpenSSL created
and the non-opaqueness of CONF.
Although we aren't likely to be supporting CONF files for the time
being, this change seems to reasonable for enough us to maintain so we
can get a better build/patch for Ruby. We still fail most CONF related
tests in Ruby, so we'll have to work around those. But this commit lets
us gracefully build with the underlying C code for Ruby CONF without any
AWS-LC specific logic.

### Testing:
Verified that this passes the build for Ruby

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.