Add missing QLANE rule to Arm decoder, fix error check in run-sematest.sh
This also fixes the grep command of run-proof.sh . `grep -r` must take a
directory name, but it was giving the output log file name.
In the case of the proof CI, it is being double-checked by printing the
list of proven specifications, so this missing check was somewhat benign.