Adding -verify and expanding -x509 options for our OpenSSL tool (#1951)
### Issues:
`CryptoAlg-2679`

### Description of changes: 
- `-subject`, `-fingerprint`, `-inform`, `-enddate`, `-subject_hash`,
`-subject_hash_old` options added to the x509 tool
- refactored the x509 tool to read from stdin
- A new `verify` tool with `-CAfile`
- The `verify` tool takes a file as input or defaults to reading from
stdin

### Callouts: 
The subject field differs in OpenSSL master and versions <= 3.2. A
normalization function that removes whitespaces is used to compare
output.

### Testing:
Added option tests for new functionality and OpenSSL comparison tests
for the new options.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.