Expose BN_set_flags as a no-op (#2021)
We're currently leaning towards continuing to not support
`BN_FLG_CONSTTIME`. It's a problematic flag that's poorly documented and
could lead to constant time violations if we aren't careful.
We can expose `BN_set_flags` as a no-op however. Ruby has logic to get
around the `BN_FLG_CONSTTIME` flag, but not around `BN_set/get_flags`.
Exposing the set symbol would allow for a cleaner patch/build.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.