Avoid access to stack pointer + negative offset in p384_montjadd
This is a patch that fixes p384_montjadd and p384_montjscalarmul (which
has p384_montjadd inlined) to avoid memory access to stack pointer +
negative offset in the middle of the function.
This is because when a process is interrupted, in principle, context
switching to a handler may overwrite data under the stack pointer:
https://forum.osdev.org/viewtopic.php?t=21720
This also caused valgrind from AWS-LC side to fail.

This is porting the patch of @dkostic:
https://github.com/aws/aws-lc/pull/2036/commits/65808f80b26171e2d66347b84f99c1acb26efee6
.. and also updates the proof according to it.
Also, this adds valgrind to the test so that it can print such cases.

s2n-bignum original commit: https://github.com/awslabs/s2n-bignum/commit/99ff4f8a84bc7ddb74e104d8ea7d59b83da68e5f