Implement PKCS7_verify, update PKCS7_sign (#1993)
This PR implements `PKCS7_verify` for verifying signedData-type PKCS7 messages.
It also adds another mode of operation to `PKCS7_sign` for compatibility with
Ruby. Flags for these functions are documented in header comments.

We also remove most support for signedAndEnveloped-type PKCS7 messges, as
they're not required by Ruby's tests beyond a few getters/setters.  OpenSSL
supports this type in [`PKCS7_decrypt`][1] but not [`PKCS7_verify`][2],
[`PKCS7_sign`][3], nor [`PKCS7_encrypt`][4], so it's unclear what (if any) role
signedAndEnveloped types have with respect to these functions. If we wish to
revisit support for this type in the future, I have a branch [here][5]
implementing encrypt/decrypt/sign/verify support and [a test][6].

Lastly, we externalize some of the previously internal PKCS7 ASN.1/struct
definitions as ruby's source accesses some of their members directly.

[1]: https://github.com/openssl/openssl/blob/f7c4d4519cd5daad5335b39ba7a3c864c65bf6ee/crypto/pkcs7/pk7_smime.c#L490
[2]: https://github.com/openssl/openssl/blob/f7c4d4519cd5daad5335b39ba7a3c864c65bf6ee/crypto/pkcs7/pk7_smime.c#L236
[3]: https://github.com/openssl/openssl/blob/f7c4d4519cd5daad5335b39ba7a3c864c65bf6ee/crypto/pkcs7/pk7_smime.c#L36
[4]: https://github.com/openssl/openssl/blob/f7c4d4519cd5daad5335b39ba7a3c864c65bf6ee/crypto/pkcs7/pk7_smime.c#L442
[5]: https://github.com/WillChilds-Klein/aws-lc/tree/pkcs7-signed-and-enveloped/
[6]: https://github.com/WillChilds-Klein/aws-lc/blob/pkcs7-signed-and-enveloped/crypto/pkcs7/pkcs7_test.cc#L1953-L1984