Remove extra va_end in err_add_error_vdata
We should va_end(args_copy), but the function should not va_end(args).
The caller is responsible for that one.

As I understand it, this in benign: in every target we care about,
va_end is a no-op.

Bug: 379096238
Change-Id: I063c775ecd496edc1b5c69c027d1dd9a04f578d1
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/74088
Auto-Submit: David Benjamin <davidben@google.com>
Reviewed-by: Adam Langley <agl@google.com>
Commit-Queue: Adam Langley <agl@google.com>