Use SHA256 as default digest for OCSP signing (#2038)
OpenSSL apparently sets a default digest when consuming the
`EVP_DigestSign*` APIs. We avoid this and emit a relevant error when
this happens. Ruby 3.2's double downed on this on their consumption
of OCSP however: ruby/ruby@4d6a293 
We don't want to introduce a "default digest" to all calls to
`EVP_DigestSign*`, so I've chosen to imitate OpenSSL's behavior in only
the places necessary in this case (OCSP). This fixes the Ruby 3.2 OCSP
tests we were failing before.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.