Update BoringSSL benchmark to use C++17 (#2063)
### Description of changes: 
BoringSSL now requires C++17 after
https://github.com/google/boringssl/commit/9ff849165a4b6c33efc2e1b10b71400ce2257fed.
When we build speed.cc with BoringSSL's headers we need to specify our
our build to use C++17. This is blocking all of our CI with failures
like
[this](https://us-west-2.codebuild.aws.amazon.com/project/eyJlbmNyeXB0ZWREYXRhIjoiWk5IUWJGRGxBcTJ2Mkp4WGF3dnBwYjc5V0ZZYSt5SVVGbkwvODkydTNTaVQ2V2FMN3hwa0tjSWNFemw2QWtCWW5welFWV3lpRFpKVitwejgvelhpRWh3NDNqcWhKalpPYW9hL2tLMDlJSDFPT1NkNyIsIml2UGFyYW1ldGVyU3BlYyI6Ilp4VXRNQXFGM1BZYVlaRkIiLCJtYXRlcmlhbFNldFNlcmlhbCI6MX0%3D/build/73b9a032-23e0-493c-b4f2-08fce7df7798).

### Call-outs:
This does not change AWS-LC's normal C++11 requirement.

### Testing:
The CI will build this. 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.