Remove algorithms from testmodulewrapper that are now used in the real modulewrapper (#2069)
### Description of changes: 
The testmodulewrapper is used to test ACVP things that the real
modulewrapper doesn't support due to lack of support in libcrypto. We
have added
[XTS](https://github.com/aws/aws-lc/blob/main/util/fipstools/acvp/modulewrapper/modulewrapper.cc#L314-L322)
and
[KDF](https://github.com/aws/aws-lc/blob/main/util/fipstools/acvp/modulewrapper/modulewrapper.cc#L1152-L1195)
to the modulewrapper and no longer need to test them here.

### Testing:
Removed code, existing tests pass.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.