FreeBSD and OpenBSD fork detection (#2089)
Two things are happening in this PR:

Refactor code in fork_detect.c to make is more readable and amendable
Add new sufficient method for fork detection on FreeBSD and OpenBSD.
The new method uses minherit, tagging a memory page with a special value. The semantics are equivalent to the existing case with madvise.