Fixes varios issues with rebuilding CI Docker images (#2077)
* Update Intel SDE from `9.21.1` to `9.44.0` as the older version is no
longer available on at the mirror location.
* Update the AL2023 SSM Agent file to pull the changed path location for
the manifest for GitHub.
* Update the CentOS 7 docker image's mirrors to use the vault location
as the previous mirror is no longer serving the repository files.
* Since cryptofuzz is no longer available on GitHub from the original
upstream author, update it to pull an archived file from a CloudFront
distribution from our team AWS account. I have also put a copy of the
code in AWS CodeCommit for a complete git history.
* Botan repository tip isn't building for us on AL2023 anymore (looks
like C++ 20 related / issues with cryptofuzz maybe?), anyways pinned to
the current cryptofuzz we've been using and botan version that was
released around that time.
* Update CDK dependencies for good hygiene
* Turns out our Windows CDK image build was never truly automated....so
fix that by automating the long-standing issue of not installing the
Docker engine.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.