Add x509-limbo patch and reporting tool (#2049)
### Description of changes: 
Describe AWS-LC’s current behavior and how your code changes that
behavior. If there are no issues this pr is resolving, explain why this
change is necessary.

### Call-outs:
* The majority of the x509-limbo patch is a 1-1 copy of the code from
the
[openssl](https://github.com/C2SP/x509-limbo/tree/main/harness/openssl)
harness code from that repository. It just has a very limited set of
changes to make it compatible with AWS-LC. I'd suggest not wasting your
effort into reviewing that code specifically.
* The Go reporting tool and patch will be built and tested in this PR,
but we won't be producing the report output and capturing it / diffing
it with the tool until a follow-up PR. I've mostly broken this up to
alleviate a large code review for everyone.
* My current intention is to wire this up into a separate CodeBuild
project similar to fuzzing. But for now (since that requires CDK work
etc), I have wired-up some simple verification of the patching and the
reporting tool into the existing ~Linux CodeBuild project~ GitHub
Actions CI. This is only temporary.

### Testing:
* I've added plenty of testing to the Go program itself for the
functionality the CLI tool relies on for how to produce the reports.
* Verified the actual CLI commands locally as well.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.

---------

Co-authored-by: Justin W Smith <103147162+justsmth@users.noreply.github.com>