Align BN_bn2hex behavior with OpenSSL (#2122)
OpenSSL's historically printed `BN_bn2hex` with upper case letters, as
opposed to our original upstream predecessors. We haven't had a concrete
ask to change this until recently:
https://github.com/ruby/openssl/issues/833.
As most of the open source community and our integration targets still
depend on OpenSSL as their main libcrypto dependency, we should consider
making the minor adjustment to minimize churn for us and our consumers.

### Testing:
Tweaks to existing tests.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.