Only need libunwind for testing (#2131)
### Notice
This change does not modify the FIPS module boundary. This change only
avoids having CMake check for the availability of `Libunwind` when no
tests are being built.

### Description of changes:
* Libunwind is only needed when building tests.
* See same change made for main: https://github.com/aws/aws-lc/pull/2093

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.