Enable RDRAND for all x86_64 except known prohibitively slow CPU models (#2135)
Remove the need to have a "fast" rdrand. Instead flip the model to exclude known prohibitively slow rdrand implementations. In particular, this includes the AMD Zen1 and Zen2 generation microarchitectures. See q/VxC3AiwXpAjJ for some data on that.

This change is made to increase the usage of prediction resistance.