Check for MIPSEB in target.h (#2143)
### Issues:
Addresses: 
* aws/aws-lc-rs#522

### Description of changes: 
Setup macros for big-endian MIPS in "target.h"

### Callout
A contributor informs us that the build for `mips-unknown-linux-musl`
succeeds with this change:
https://github.com/aws/aws-lc-rs/issues/522#issuecomment-2618841038

### Testing:
No CI testing for it. 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.