Fix issue with ML-DSA key parsing (#2152)
This fixes a memory read misalignment in the ML-DSA public and private
key parsing. I've also added new tests to verify that the test works.
This should also fix the intermittent errors we've been having in our
fuzz tests. The plan is to cut a release as soon as this is merged.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.