Add support for PKCS12_set_mac (#2128)
 Ruby has added a binding for an additional function called
`PKCS12_set_mac` in it's master branch: ruby/ruby@c79b435.
This was discovered prior to the PR for Ruby's CI integration with the
master branch being merged: https://github.com/aws/aws-lc/pull/2071

OpenSSL's implementation of the function directly sets a designated mac
field within the `PKCS12` structure. Our `PKCS12` structure is folded
into a string of bytes along with its length and there aren't any
available fields for us to directly set. This means that AWS-LC's
implementation of `PKCS12_set_mac` has to parse the proper contents from
`PKCS12` and rerun the key and mac generation with the new parameters
provided.

### Call-outs:
1. The parsing logic is similar to `PKCS12_get_key_and_certs`, but the
logic can't be shared. This is because we need to maintain pointers to
certain parts of the CBS parsers in `PKCS12_set_mac`, so that we can
properly rewrite the contents later.
2. I've abstracted the key and mac generation in `PKCS12_create` and
`PKCS12_set_mac` to a single function called `pkcs12_gen_and_write_mac`.
3. OpenSSL 3.0 uses SHA-256 as the default hash, yet OpenSSL 1.1.1 uses
SHA-1. We've decided to go with SHA-1 for now for better 1.1.1 compatibility.
HMAC's security also does not rely on the hash function's collision resistance
property.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.