Flip to new randomness generation backend and remove the existing (#2147)
This PR removes the existing randomness generation and replaces it with the new one. It does a number of things to do that:

removes "unsafe buffering" mode. This is not a thing in the new implementation and was never used.
move rdrand implementation to the entropy source implementation file
removes the ability to specify the "entropy source" at compile time. The flag is retained for build backwards compatibility. There is only a single (validable) entropy source left, and that is jitter entropy. (other sources are used in the randomness generation workflow, but not validatable).
passive entropy is removed completely.
the PRNG model test is disabled until I fix it.
RAND_bytes_with_user_prediction_resistance() is introduced as a public API. It replaces the internal RAND_bytes_with_additional_data().
correct sandbox test to account for getrandom usage.
disable state validation in rand_bytes_core when snapsafe testing is active. Otherwise, we have threading issues.