Ensure enabling local symbols doesn't change the module hash (#2169)
### Description of changes: 
The original intention of https://github.com/aws/aws-lc/pull/1005 was to
ensure break-kat.go always worked with the library.
https://github.com/aws/aws-lc/pull/1252 then used a side effect to get
the module hash but didn't actually fail the build if the hash was
different.

Turn on `pipefail` so the script exits unsuccessfully if any command
fails even in a pipe. Previously if test-break-kat.sh failed the script
didn't exit, instead it would continue to the grep which would also fail
to find any matching string, and then the `|| true` ensured the script
always continued on.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.