Add EVP API Support for ED25519ph (#2144)
### Description of changes: 
Add support for HashEdDSA (Ed25519ph) to EVP APIs with support for
setting the associated signature context. Opting not to expose
Ed25519ctx explicitly at this time since there hasn't been an ask for
it. Worse case scenario a user desiring it will have access to the
low-level APIs where it is available.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.